refactor(*): move content to stdlib
Almost everything in this repository was moved to the stdlib, so now it is basically empty. @johoelzl I recommend moving the real numbers into a branch of stdlib. We will continue to use this repository for experimental work, but most material should end up in stdlib for the time being.